test: add n8n import test modules for expressions, graph, and node mapping
Adds four new test modules covering the n8n import pipeline: expression parsing, graph construction, node mapping, and a general test harness. These tests validate the core import logic and ensure correctness of the n8n workflow conversion.

Auto-committed-on: dragonfly
